Rename collision_detection to collisions
