+++ doc/flame/collision_detection.md
@@ -208,8 +208,9 @@ implementation is performant enough for you, you probably don't have to read thi
 A broad phase is the first step of collision detection where potential collisions are calculated.
 To calculate these potential collisions are a lot cheaper to calculate than to check the exact
 intersections from the directly and it removes the need to check all hitboxes against each other
-and therefore avoiding O(n²). The set of these potential collisions are then used to check the
-exact intersections between hitboxes, this is sometimes called narrow phase.
+and therefore avoiding O(n²). The broad phase produces a set of potential collisions (a set of 
+`CollisionProspect`s), this set is then used to check the exact intersections between hitboxes, this
+is sometimes called narrow phase.
 
 By default Flame's collision detection is using a sweep and prune broadphase step, if your game
 requires another type of broadphase you can write your own broadphase by extending `Broadphase` and
