+++ examples/lib/stories/collision_detection/circles_example.dart
@@ -1,4 +1,4 @@
-import 'package:flame/collision_detection.dart';
+import 'package:flame/collisions.dart';
 import 'package:flame/components.dart';
 import 'package:flame/game.dart';
 import 'package:flame/input.dart';
