+++ examples/lib/stories/collision_detection/multiple_shapes_example.dart
@@ -1,6 +1,6 @@
 import 'dart:math';
 
-import 'package:flame/collision_detection.dart';
+import 'package:flame/collisions.dart';
 import 'package:flame/components.dart';
 import 'package:flame/extensions.dart';
 import 'package:flame/game.dart';
