--- packages/flame/lib/collision_detection.dart
@@ -1,11 +1,0 @@
-export 'src/collision_detection/broadphase.dart';
-export 'src/collision_detection/collision_callbacks.dart';
-export 'src/collision_detection/collision_detection.dart';
-export 'src/collision_detection/collision_passthrough.dart';
-export 'src/collision_detection/hitboxes/hitbox.dart';
-export 'src/collision_detection/hitboxes/hitbox_circle.dart';
-export 'src/collision_detection/hitboxes/hitbox_polygon.dart';
-export 'src/collision_detection/hitboxes/hitbox_rectangle.dart';
-export 'src/collision_detection/hitboxes/hitbox_shape.dart';
-export 'src/collision_detection/sweep.dart';
-export 'src/components/screen_collidable.dart';
