+++ packages/flame/lib/collisions.dart
@@ -1,0 +1,11 @@
+export 'src/collisions/broadphase.dart';
+export 'src/collisions/collision_callbacks.dart';
+export 'src/collisions/collision_detection.dart';
+export 'src/collisions/collision_passthrough.dart';
+export 'src/collisions/hitboxes/hitbox.dart';
+export 'src/collisions/hitboxes/hitbox_circle.dart';
+export 'src/collisions/hitboxes/hitbox_polygon.dart';
+export 'src/collisions/hitboxes/hitbox_rectangle.dart';
+export 'src/collisions/hitboxes/hitbox_shape.dart';
+export 'src/collisions/sweep.dart';
+export 'src/components/screen_collidable.dart';
