+++ packages/flame/lib/src/collisions/collision_callbacks.dart
