+++ packages/flame/test/collision_detection/collidable_type_test.dart
@@ -1,6 +1,6 @@
 import 'dart:math';
 
-import 'package:flame/collision_detection.dart';
+import 'package:flame/collisions.dart';
 import 'package:flame/components.dart';
 import 'package:flame/game.dart';
 import 'package:flame_test/flame_test.dart';
