+++ packages/flame/test/collision_detection/collision_callback_benchmark.dart
@@ -1,6 +1,6 @@
 import 'dart:math';
 
-import 'package:flame/collision_detection.dart';
+import 'package:flame/collisions.dart';
 import 'package:flame/components.dart';
 import 'package:flame_test/flame_test.dart';
 import 'package:test/test.dart';
