+++ packages/flame/test/collision_detection/collision_test_helpers.dart
@@ -1,4 +1,4 @@
-import 'package:flame/collision_detection.dart';
+import 'package:flame/collisions.dart';
 import 'package:flame/components.dart';
 import 'package:flame/game.dart';
 import 'package:flame_test/flame_test.dart';
