+++ packages/flame_bloc/example/lib/src/game/components/bullet.dart
@@ -1,4 +1,4 @@
-import 'package:flame/collision_detection.dart';
+import 'package:flame/collisions.dart';
 import 'package:flame/components.dart';
 
 import '../../inventory/bloc/inventory_bloc.dart';
