+++ packages/flame_bloc/example/lib/src/game/components/enemy.dart
@@ -1,4 +1,4 @@
-import 'package:flame/collision_detection.dart';
+import 'package:flame/collisions.dart';
 import 'package:flame/components.dart';
 
 import './explosion.dart';
