+++ packages/flame_bloc/example/lib/src/game/components/player.dart
@@ -1,4 +1,4 @@
-import 'package:flame/collision_detection.dart';
+import 'package:flame/collisions.dart';
 import 'package:flame/components.dart';
 import 'package:flame/input.dart';
 import 'package:flame_bloc/flame_bloc.dart';
